requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==22.2.0
chardet==4.0.0
dateutils==0.6.12
discord.py==1.7.3
idna==3.4
multidict==6.0.4
mysqlclient==2.1.1
Pillow==9.5.0
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
typing_extensions==4.5.0
yarl==1.8.2
